chore(gyms): 🚚 Create controller file for gyms
